ipnewton linesearch: switch to one-sided tests
This appears to fix many fail-to-converge problems. In some cases, the Hessian prediction of the slope change is approximately or exactly 0; in that case the slope condition is impossible to satisfy. A cure is to use the positive-definite modified Hessian (which will never give 0 for the estimated Hessian effect on the slope) in the computation of the linesearch local-model estimate. This only makes sense in the context of switching the line search slope criterion to an upper bound (which is sensible in its own right).

An unfortunate consequence is that we have to compute 3 cholesky factorizations rather than 2. The 3rd is necessary only for the linesearch, which is unfortunate.